se subio el segundo modelo funciona
creditos a Juan Martin Leoni por enseñarme a hacer el historial de operaciones
