+++ HIST.json
@@ -1,0 +1,10 @@
+[
+  {
+    "cuenta": "81000+2",
+    "resultado": 81002
+  },
+  {
+    "cuenta": "4+4",
+    "resultado": 8
+  }
+]
