--- Segundo modelo  de calcu/calcu/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "calculator",
-  "version": "1.0.0",
-  "description": "Trabajo practico",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "all": "^0.0.0",
-    "better-logging": "^5.0.0",
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "ejs": "^3.1.9",
-    "express": "^4.18.2",
-    "fs": "^0.0.1-security"
-  }
-}
