--- Segundo modelo  de calcu/calcu/src/helpers/renderView.js
@@ -1,5 +1,0 @@
-const renderView = (res, view, options = {}) => {
-    res.render('layout', { content: view, ...options});
-}
-
-module.exports = { renderView };
