--- Segundo modelo  de calcu/calcu/views/layout.ejs
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <title>TP N2 - Calculadora</title>
-
-  <!-- <link href="https://fonts.googleapis.com/css2?family=Montserrat&family=Open+Sans&display=swap" rel="stylesheet">
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/normalize.css@8.0.1/normalize.min.css">
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/glider-js@1.7.3/glider.min.css">
-
-  <script src="https://cdn.jsdelivr.net/npm/glider-js@1.7.3/glider.min.js"></script>
-  <script src="https://kit.fontawesome.com/2c36e9b7b1.js" crossorigin="anonymous"></script> -->
-
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@500&display=swap" rel="stylesheet">
-
-</head>
-
-<body>
-  <!-- <h1>Mi Aplicación</h1> -->
-  <div>
-    <!-- Aquí se insertará el contenido específico para cada vista -->
-    <%- include(`${content}`) %> 
-  </div>
-</body>
-
-</html>
