+++ package.json
@@ -12,8 +12,10 @@
   "dependencies": {
     "all": "^0.0.0",
     "better-logging": "^5.0.0",
+    "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "ejs": "^3.1.9",
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "fs": "^0.0.1-security"
   }
 }
