+++ views/layout.ejs
@@ -2,13 +2,16 @@
 <html>
 
 <head>
-  <title>TP N2 - Calculadora - NICOLAS SALDAÑA</title>
+  <title>TP N2 - Calculadora - SEGUNDO MODELO NICOLAS SALDAÑA</title>
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@500&display=swap" rel="stylesheet">
 
 </head>
 
 <body>
-  <div style="margin: auto;">
-    <%- include(`${content}`) %>
+  <div>
+    <%- include(`${content}`) %> 
   </div>
 </body>
